Refactoring the main loop
Created a table of opcodes, eliminating the big buggy switch
Moved the main loop to the frame object
Hacked the outputDevice through frame objects
